requirements.txt
# Environment variable loading (.env file support)
python-dotenv==1.0.1

# Anthropic Claude API client
anthropic==0.49.0

# YAML config file support
pyyaml==6.0.2

# HTTP requests (backup, in case we need it for non-PRAW sources)
requests==2.32.3

# RSS feed parsing (Google Alerts scraper)
feedparser==6.0.12

# Gmail API (draft creation for BID proposals)
google-api-python-client==2.166.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
